Protect certain pages and features from unauthorized users

+++ flask_app.py
@@ -140,6 +140,8 @@ def index():
 
 @app.route("/project_entry", methods=["POST", "GET"])
 def project_entry():
+    if session['account_type'] != "full":
+        return redirect(url_for("payment"))
     all_contacts_query = "SELECT * FROM contacts WHERE company_id=?"
     with create_connection() as conn:
         all_contacts = pd.read_sql(all_contacts_query, conn, params=[session.get('company_id')])
@@ -356,6 +358,8 @@ def potential_match():
 
 @app.route("/summary_table")
 def summary_table():
+    if session['account_type'] != "full":
+        return redirect(url_for("payment"))
     def highlight_pending(s):
         days_old = (
             datetime.datetime.now().date()
@@ -497,6 +501,8 @@ def summary_table():
 
 @app.route("/delete_job")
 def delete_job():
+    if session['account_type'] != "full":
+        return redirect(url_for("payment"))
     delete_job_query = """
             DELETE FROM company_projects
             WHERE project_id=?
@@ -514,6 +520,8 @@ def delete_job():
 
 @app.route("/instant_scan", methods=["POST", "GET"])
 def instant_scan():
+    if session['account_type'] != "full":
+        return redirect(url_for("payment"))
     if request.method == "POST":
         job_number = request.args.get("job_number")
         lookback = request.args.get("lookback")
@@ -590,6 +598,8 @@ def payment():
 
 @app.route("/contact_config", methods=["POST", "GET"])
 def contact_config():
+    if session['account_type'] != "full":
+        return redirect(url_for("payment"))
     contact = request.args
     all_contacts_query = "SELECT * FROM contacts WHERE company_id=?"
     with create_connection() as conn:
@@ -634,6 +644,8 @@ def contact_config():
 
 @app.route("/delete_contact")
 def delete_contact():
+    if session['account_type'] != "full":
+        return redirect(url_for("payment"))
     delete_contact_query = """
             DELETE FROM contacts
             WHERE id=?
@@ -647,6 +659,8 @@ def delete_contact():
 
 @app.route("/update_contact", methods=["POST", "GET"])
 def update_contact():
+    if session['account_type'] != "full":
+        return redirect(url_for("payment"))
     contact = request.args
     update_contact_query = """
         UPDATE contacts
@@ -666,6 +680,8 @@ def update_contact():
 
 @app.route("/add_contact", methods=["POST", "GET"])
 def add_contact():
+    if session['account_type'] != "full":
+        return redirect(url_for("payment"))
     contact = request.args
     add_contact_query = """
         INSERT INTO contacts
@@ -1081,7 +1097,8 @@ def map():
             tooltip=f"{row.title[:25]}{'...' if len(row.title) >= 25 else ''}",
             icon=folium.Icon(prefix='fa', icon='check', color='gray')
         ).add_to(feature_group)
-    feature_group.add_to(m)
+    if session['account_type'] == "full":
+        feature_group.add_to(m)
 
     feature_group = folium.FeatureGroup(name='Open Projects')
     for _, row in df_cp_open.iterrows():
@@ -1134,7 +1151,8 @@ def map():
             tooltip=f"{row.title[:25]}{'...' if len(row.title) >= 25 else ''}",
             icon=folium.Icon(prefix='fa', icon='search', color='black')
         ).add_to(feature_group)
-    feature_group.add_to(m)
+    if session['account_type'] == "full":
+        feature_group.add_to(m)
 
     feature_group = folium.FeatureGroup(name="Web CSP's")
     for _, row in df_wc.iterrows():
